// src/enums/analytics.enum.ts
/** Datatype enum for Custom Firebase Events */
export enum CustomFirebaseEvents {
	SCREEN_VIEW = "screen_view",
	USER_SIGN_UP_ATTEMPT = "user_sign_up_attempt",
	USER_LOGIN_ATTEMPT = "user_login_attempt",
	USER_LOGIN_SUCCESS = "user_login_success",
	USER_SIGN_UP_SUCCESS = "user_sign_up_success",
	USER_LOGIN_FAILURE = "user_login_failure",
	USER_SIGN_UP_FAILURE = "user_sign_up_failure",
	USER_APP_OPENED = "user_app_opened",
	USER_LOGOUT = "user_logout",
	FORGOT_PASSWORD = "forgot_password",
	RESET_PASSWORD = "reset_password",
}